requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
Django==5.2.7
flatbuffers==25.9.23
humanfriendly==10.0
idna==3.11
imageio==2.37.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
protobuf==6.33.0
PyMatting==1.1.14
pyreadline3==3.5.4
referencing==0.37.0
rembg==2.0.67
requests==2.32.5
rpds-py==0.28.0
scikit-image==0.25.2
scipy==1.16.2
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.10.16
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
